Fixed compilation errors.

+++ include/dmtcp.h
@@ -156,7 +156,7 @@ EXTERNC int dmtcp_enable_ckpt(VOID) __attribute__ ((weak));
 #define dmtcp_enable_ckpt() \
  (dmtcp_enable_ckpt ? dmtcp_enable_ckpt() : DMTCP_NOT_PRESENT)
 
-// See: test/plugin/sleep1 dir and sibling directories for examples: 
+// See: test/plugin/sleep1 dir and sibling directories for examples:
 EXTERNC void dmtcp_event_hook(DmtcpEvent_t event, DmtcpEventData_t *data)
   __attribute((weak));
 
@@ -244,7 +244,7 @@ EXTERNC int dmtcp_get_local_status(int *numCheckpoints, int *numRestarts)
 EXTERNC int dmtcp_is_running_state(void);
 // Primarily for use by the modify-env plugin.
 EXTERNC int dmtcp_get_restart_env(const char *name,
-                                  char *value, int maxvaluelen);
+                                  char *value, size_t maxvaluelen);
 // Get pathname of target executable under DMTCP control.
 EXTERNC const char* dmtcp_get_executable_path();
 // True if dmtcp_launch called with --no-coordinator
